% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segunda_edicion.R
\docType{data}
\name{segunda_edicion}
\alias{segunda_edicion}
\title{Datos normativos de la segunda edición del IAPS en argentina}
\format{
Una base de datos con 1236 observaciones y 10 variables:
\describe{
\item{set}{Número de set al que pertenece el estímulo}
\item{numero}{Número de id del estímulo}
\item{desc}{Descripción breve del contenido del estímulo}
\item{grupo}{Grupo al que pertece cada observación. Puede ser población general, grupo de mujeres o grupo de hombres}
\item{val_M}{Variable Valencia: Valor medio de la agradabilidad que generó el estímulo}
\item{val_DE}{Desviación típica para la variable Valencia}
\item{aro_M}{Variable Arousal/Activación: Valor medio de la activación que generó el estímulo}
\item{aro_DE}{Desviación típica para la variable Valencia}
\item{con_M}{Variable Control: Valor medio de la sensación de control que generó el estímulo}
\item{con_DE}{Desviación típica para la variable Control}
}
}
\source{
\url{https://doi.org/10.32348/1852.4206.v12.n3.24471}
}
\usage{
segunda_edicion
}
\description{
Datos normativos de la segunda edición del IAPS (Lang et al., 2008) una base de estímulos para inducción emocional en el laboratorio
}
\keyword{datasets}
